separate participants schema
